Clean Code bei employeeManagementView.kt, parameterview.kt und PopUps.
