+++ 1_Semesterarbeit/Ressourcix/kotlin/ressourcix/gui/pages/parameterview.kt
@@ -4,28 +4,19 @@ package ressourcix.gui.pages
 
 import javafx.geometry.Insets
 import javafx.geometry.Pos
-import javafx.scene.Node
-import javafx.scene.control.Button
-import javafx.scene.control.Label
 import javafx.scene.control.TextField
-import javafx.scene.control.TextFormatter
 import javafx.scene.layout.BorderPane
 import javafx.scene.layout.HBox
-import javafx.scene.layout.Region
-import javafx.scene.layout.StackPane
 import javafx.scene.layout.VBox
-import javafx.scene.text.TextAlignment
 import javafx.scene.input.KeyEvent
 import ressourcix.domain.config
 import ressourcix.gui.popups.weekPickerPopUp
 import ressourcix.logger.logger
-
-private const val BTN_HEIGHT = 50.0
-private const val BTN_WIDTH = 140.0
-private const val TFL_HEIGHT = 30.0
-private const val TFL_WIDTH = 300.0
-
-//TODO Clean Code bearbeitung mit employeemanagementView.kt
+import ressourcix.gui.util.createButton
+import ressourcix.gui.util.createDataBox
+import ressourcix.gui.util.createTfl
+import ressourcix.gui.util.positiveIntNoZeroFormatter
+import ressourcix.gui.components.PopupLayer
 
 object parameterview: BorderPane() {
 
@@ -82,34 +73,12 @@ object parameterview: BorderPane() {
             logger.info("Parameter wurden erfolgreich zurückgesetzt.")}
     }
 
-    private val dim = Region().apply {
-        style = "-fx-background-color: rgba(0,0,0,0.35);"
-        isVisible = false
-        isMouseTransparent = false
-        isManaged = true
-        setOnMouseClicked { closePopup() }
-    }
-
-    private val popupHost = StackPane().apply {
-        isVisible = false
-        isMouseTransparent = false
-        isManaged = true
-        alignment = Pos.CENTER
-        maxWidth = Double.MAX_VALUE
-        maxHeight = Double.MAX_VALUE
-
-    }
+    private val popupLayer = PopupLayer()
 
     private val mainContent = VBox().apply {
         padding = Insets(50.0)
         alignment = Pos.CENTER
         spacing = 50.0
-    }
-
-    private val centerStack = StackPane().apply {
-        children.addAll(mainContent, dim, popupHost)
-        StackPane.setAlignment(dim, Pos.CENTER)
-        StackPane.setAlignment(popupHost, Pos.CENTER)
     }
 
     private val minEmployeeBar = HBox(50.0).apply {
@@ -156,60 +125,10 @@ object parameterview: BorderPane() {
 
     init {
         mainContent.children.addAll(minEmployeeBar,vacationsBlockerBar,maxVacationsBar,functionBox)
-        center = centerStack
+        center = popupLayer.wrap(mainContent) { popupLayer.close() }
         setFieldsEditable(false)
         installWeekPopupHandlers()
         loadFromConfigToUi()
-    }
-
-
-    fun showPopup(popupContent: Node) {
-        popupHost.children.setAll(popupContent)
-        dim.isVisible = true
-        popupHost.isVisible = true
-        dim.toFront()
-        popupHost.toFront()
-    }
-
-    fun closePopup() {
-        popupHost.children.clear()
-        dim.isVisible = false
-        popupHost.isVisible = false
-    }
-
-    private fun createTfl(text: String,prompt: String): TextField =
-        TextField(text).apply {
-            promptText = prompt
-            prefHeight = TFL_HEIGHT
-            prefWidth = TFL_WIDTH
-            isFocusTraversable = false
-        }
-
-    private fun createButton(text: String): Button =
-        Button(text).apply {
-            prefHeight = BTN_HEIGHT
-            prefWidth = BTN_WIDTH
-            textAlignment = TextAlignment.CENTER
-            alignment = Pos.CENTER
-            isFocusTraversable = false
-            style = " -fx-font-weight: bold;"
-        }
-
-    private fun createDataBox(labelText: String, field: TextField): VBox =
-        VBox(6.0).apply {
-            children.addAll(Label(labelText).apply {
-                style = " -fx-font-weight: bold;"
-            }, field)
-        }
-
-    private fun positiveIntNoZeroFormatter(maxDigits: Int,maxValue: Int): TextFormatter<String> {
-        return TextFormatter { change ->
-            val newText = change.controlNewText
-            val ok = newText.isEmpty() || (newText.matches(Regex("[1-9][0-9]*"))
-                    && newText.length <= maxDigits
-                    && newText.toInt() <= maxValue)
-            if (ok) change else null
-        }
     }
 
     private fun installWeekPopupHandlers() {
@@ -220,9 +139,9 @@ object parameterview: BorderPane() {
                     "Ferien Blocker",
                     weeks = weeksForBlocker,
                     onTextChanged = { txt -> vacationsBlockTfl.text = txt },
-                    onClose = { closePopup() }
+                    onClose = { popupLayer.close() }
                 )
-                showPopup(popup.build())
+                popupLayer.show(popup.build())
                 e.consume()
             }
         }
@@ -233,9 +152,9 @@ object parameterview: BorderPane() {
                     "Schulferien",
                     weeks = weeksForSchool,
                     onTextChanged = { txt -> vacationsSchoolBlockTfl.text = txt },
-                    onClose = { closePopup() }
+                    onClose = { popupLayer.close() }
                 )
-                showPopup(popup.build())
+                popupLayer.show(popup.build())
                 e.consume()
             }
         }
@@ -282,7 +201,6 @@ object parameterview: BorderPane() {
             if (s == e) "KW $s" else "KW $s-$e"
         }
     }
-
 
     private fun loadFromConfigToUi() {
         minEmployeeTfl.text = config.minEmployeeNumber.toString()
