+++ 1_Semesterarbeit/Ressourcix/kotlin/ressourcix/gui/util/uiConstants.kt
@@ -1,0 +1,6 @@
+package ressourcix.gui.util
+
+const val BTN_HEIGHT = 50.0
+const val BTN_WIDTH = 140.0
+const val TFL_HEIGHT = 30.0
+const val TFL_WIDTH = 300.0
